requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
curl_cffi==0.13.0
distro==1.9.0
Django==5.2.3
django-cors-headers==4.9.0
django-cron==0.6.0
django-pwa==2.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multitasking==0.0.12
mysqlclient==2.2.7
numpy==2.3.2
openai==1.98.0
openpyxl==3.1.5
pandas==2.3.2
peewee==3.18.2
platformdirs==4.3.8
protobuf==6.32.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.11.0
yfinance==0.2.65
